Add tests for clients and redirect URIs.
Add tests for clients and their redirect URIs, and fix the problems the
tests exposed.

Update the error returned when a redirect URI is missing to be more
human-friendly.

Remove unreachable error checks in ClientService methods.

Update typos in handling code for redirect URIs, both in the response
body key and in the URLs for API calls.