Better User-Agent support.
Add more sophisticated handling of user agents, allowing appending and
prepending to the user agent.

Add a mutex around modifying the client credentials, to protect against
race conditions.

Fix a bug where HMAC requests had their body read by the HMAC signing
code and didn't get reset so the server could also read them. Instead,
read them into a buffer, sign the buffer, and replace the body with the
buffer.

Fix bug in test code that would inspect the request and not reset it for
further inspection or use. Use a new cloneRequest helper method to make
copies of requests so they can be read multiple times.

Add a test helper to make sure a request was authorized using HMAC.

Add a test for the user-agent behavior that we added.